_posts_pandoc/requirements.txt
python_frontmatter>=0.5.0
